Added timeout in the connecting to Teensy Changed hfi_abs_pos to double Adjusted rel setpoint option to wait without getsig (such that tracebg can be running)

+++ Firmware/defines.h
@@ -222,7 +222,7 @@ typedef struct mot_state_t {
   float delta_id;
   float delta_iq;
   bool hfi_firstcycle;
-  float hfi_abs_pos;
+  double hfi_abs_pos;
   bool hfi_useforfeedback;
   bool hfi_use_lowpass;
   float hfi_prev;
